Rename bench classics
